requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.4
    # via requests
confluent-kafka==2.12.0
    # via -r requirements.in
delta-spark==4.0.0
    # via -r requirements.in
idna==3.11
    # via requests
importlib-metadata==8.7.0
    # via delta-spark
py4j==0.10.9.9
    # via pyspark
pyspark==4.0.1
    # via
    #   -r requirements.in
    #   delta-spark
pyyaml==6.0.3
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
tomli==2.3.0
    # via confluent-kafka
urllib3==2.5.0
    # via requests
zipp==3.23.0
    # via importlib-metadata
